move constants to own file and photo-handling code to separate file
